requirements.txt
appnope==0.1.4
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.14.0
fonttools==4.52.1
fsspec==2024.5.0
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.3
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.5.0
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mpmath==1.3.0
nest_asyncio==1.6.0
networkx==3.3
numpy==1.26.4
onedrivedownloader==1.1.3
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
platformdirs==4.2.2
prompt-toolkit==3.0.42
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0
pytz==2024.1
pyzmq==25.1.2
requests==2.32.2
scikit-learn==1.5.0
scipy==1.13.1
setuptools==69.5.1
six==1.16.0
stack-data==0.6.2
sympy==1.12
threadpoolctl==3.5.0
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tornado==6.3.3
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
wheel==0.43.0
zipp==3.17.0
